+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog of RorVsWild
 
+* Print ASCII server timing in logs
+
 ## 1.7.0 (2024-04-22)
 
 * Sample requests and jobs to lower our customers' bills